Fixed stuff
So... when I downloaded this script, it did not work, and now it does.
If you still care about this I suggest you test it through again. When I
downloaded this script, I encountered the following errors:
* Download dir could not be created, mkdir() was not being called
correctly
* Curl tried to print the entire stream onto the terminal in UTF-8
* The progress percentage was broken and it spammed itself until the
current file was done
I sort of fixed all of them, loosely, but you still have to config the
download dir.